Removed the global namespace std in mud.h.
That was a slog. Practically every file had hidden std:: things in it.

+++ src/pfiles.h
@@ -29,4 +29,4 @@
 #pragma once
 
 extern std::chrono::system_clock::time_point new_pfile_time_t;
-extern short num_pfiles;
+extern int num_pfiles;
